requirements.txt
matplotlib>=3.7
mss>=9.0
Pillow>=10.0
psutil>=5.9
pywin32>=306
sounddevice>=0.4.6
soundfile>=0.12
